Move settings to its own class
